requirement.txt
face_alignment
torch
Pillow
torchvision
tensorboardx